Add js code for javasript translation
Normally this should be in a single twig file and defined in the template php.
Unfortunatly the template is not defined in the fom module but in the MBCoreBundle.
To avoid further cross referencing we defined it here and will remove it later when we
refactor the complete FOM and Mapbender structure.