Replace inline twig script code with single file
This code was defined inline in the template. This is not a good practice becasue we should try to separate markup and logic.
Fix wrong ajax action path to be relative from actual page and not from the root. Change translation to js logic.
Change markup generation to jQuery for better readabilty.