Implement __toString-Method for FOM/User
Some parts of the Symofony User machinery are not strict on the type of some User-Objects. If we try to implement other login mechanics instead of form_login based on the FOM/User it can happen that we get a User-Object in the loadUserByUsername-Method if a Symfomy UserProvider. This method expects a String as an argument, therefore we need this __toString-Methods. Otherwise we get a cast-Execption.